Revert "feat: strf-10437 Autofix script: conditional file import (#1064)"
This reverts commit d00f35c3d7c99314cc1f05aea7c5c12b682ebbf9.
