+++ .github/workflows/pull_request_review.yml
@@ -11,7 +11,7 @@ jobs:
     build:
         strategy:
             matrix:
-                node: [16.x, 18.x]
+                node: [14.x]
                 os: ['ubuntu-latest', 'windows-2019', 'macos-latest']
 
         runs-on: ${{ matrix.os }}
