+++ .github/workflows/release.yml
@@ -11,7 +11,7 @@ jobs:
             - uses: actions/checkout@v3
             - uses: actions/setup-node@v3
               with:
-                  node-version: '18.x'
+                  node-version: '14.x'
             - run: npm i
             - name: Check Git Commit name
               run: git log -1 --pretty=format:"%s" |  npx commitlint
