+++ .nvmrc
@@ -1,1 +1,1 @@
-v18.13
+v14.19
