+++ lib/StencilCLISettings.js
@@ -13,6 +13,14 @@ class StencilCLISettings {
     isVerbose() {
         return this.versboseNetworkLogging;
     }
+
+    useOldNodeSassFork(flag) {
+        this.oldNodeSassFork = flag;
+    }
+
+    isOldNodeSassForkEnabled() {
+        return this.oldNodeSassFork;
+    }
 }
 
 const settings = new StencilCLISettings();
