+++ lib/bundle-validator.spec.js
@@ -3,6 +3,7 @@ const { promisify } = require('util');
 
 const ThemeConfig = require('./theme-config');
 const BundleValidator = require('./bundle-validator');
+const stencilCLISettings = require('./StencilCLISettings');
 
 const themePath = path.join(process.cwd(), 'test/_mocks/themes/valid');
 
@@ -15,6 +16,7 @@ describe('BundleValidator', () => {
 
     afterEach(() => {
         jest.restoreAllMocks();
+        stencilCLISettings.useOldNodeSassFork(false);
     });
 
     it('should not run image validations for private themes', async () => {
