+++ lib/cliCommon.js
@@ -98,6 +98,7 @@ function applyCommonOptions(program) {
 
 function setupCLI(options) {
     stencilCLISettings.setVerbose(options.verbose);
+    stencilCLISettings.useOldNodeSassFork(options.useOldNodeSassFork);
 }
 
 function prepareCommand(program) {
