+++ lib/css/compile.js
@@ -4,6 +4,7 @@ const { promisify } = require('util');
 
 const cssAssembler = require('../css-assembler');
 
+const FALLBACK_SASS_ENGINE_NAME = 'node-sass-fork';
 const SASS_ENGINE_NAME = 'node-sass';
 
 const compile = async (configuration, themeAssetsPath, fileName, engineName = SASS_ENGINE_NAME) => {
@@ -28,7 +29,7 @@ const compile = async (configuration, themeAssetsPath, fileName, engineName = SA
         sourceMap: true,
         autoprefixerOptions: {
             cascade: configuration.autoprefixer_cascade,
-            overrideBrowserslist: configuration.autoprefixer_browsers,
+            browsers: configuration.autoprefixer_browsers,
         },
     };
     const stencilStyles = new StencilStyles(console);
@@ -40,4 +41,5 @@ const compile = async (configuration, themeAssetsPath, fileName, engineName = SA
 module.exports = {
     compile,
     SASS_ENGINE_NAME,
+    FALLBACK_SASS_ENGINE_NAME,
 };
