+++ package.json
@@ -1,11 +1,11 @@
 {
   "name": "@bigcommerce/stencil-cli",
-  "version": "7.0.0-beta",
+  "version": "6.0.0",
   "description": "CLI tool to run BigCommerce Stores locally for theme development.",
   "main": "index.js",
   "engines": {
-    "node": ">=16.0.0 <19.0.0",
-    "npm": ">=8.0.0 <10.0.0"
+    "node": ">=14.0.0 <15.0.0",
+    "npm": ">=v6.11.3 <9.0.0"
   },
   "scripts": {
     "lint": "eslint . && prettier --check . --ignore-unknown",
@@ -23,8 +23,7 @@
     "stencil-pull": "./bin/stencil-pull.js",
     "stencil-start": "./bin/stencil-start.js",
     "stencil-release": "./bin/stencil-release.js",
-    "stencil-debug": "./bin/stencil-debug.js",
-    "stencil-scss-autofix": "./bin/stencil-scss-autofix.js"
+    "stencil-debug": "./bin/stencil-debug.js"
   },
   "config": {
     "stencil_version": "1.0"
@@ -49,7 +48,7 @@
   ],
   "dependencies": {
     "@bigcommerce/stencil-paper": "4.2.1",
-    "@bigcommerce/stencil-styles": "5.1.0",
+    "@bigcommerce/stencil-styles": "4.0.0",
     "@hapi/boom": "^10.0.0",
     "@hapi/glue": "^8.0.0",
     "@hapi/h2o2": "^9.1.0",
@@ -78,8 +77,6 @@
     "npm-which": "^3.0.1",
     "object-to-spawn-args": "^2.0.0",
     "parse-json": "^5.2.0",
-    "postcss": "^8.4.21",
-    "postcss-scss": "^4.0.6",
     "progress": "^2.0.3",
     "recursive-readdir": "^2.2.2",
     "semver": "^7.3.2",
