--- test/_mocks/themes/invalid-scss-latest-node-sass-to-fix/assets/scss/test.scss
@@ -1,3 +1,0 @@
-h1 {
-    color: #000000;
-}
