--- test/_mocks/themes/invalid-scss-latest-node-sass-to-fix/assets/scss/theme.scss
@@ -1,9 +1,0 @@
-h1 {
-    color: #0000AA;
-}
-
-$font: "Arial";
-
-@if not contains($font, "'Clear Sans', sans-serif") {
-    @import "test";
-}
