--- test/_mocks/themes/invalid-scss-latest-node-sass-to-fix/config.json
@@ -1,106 +1,0 @@
-{
-  "name": "Stencil",
-  "version": "1.0",
-  "css_compiler": "scss",
-  "autoprefixer_cascade": true,
-  "autoprefixer_browsers": [
-    "> 5% in US"
-  ],
-  "meta": {
-    "price": 0,
-    "documentation_url": "https://stencil.bigcommerce.com/docs/what-is-stencil",
-    "author_name": "BigCommerce",
-    "author_email": "support@bigcommerce.com",
-    "author_support_url": "https://www.bigcommerce.com/support",
-    "composed_image": "composed.jpg",
-    "features": [
-      "fully_responsive"
-    ]
-  },
-  "settings": {
-    "color": "#ffffff",
-    "font": "Sans Something",
-    "select": "first",
-    "checkbox": true,
-    "radio": "first"
-  },
-  "images": {
-    "logo": {
-      "width": 100,
-      "height": 100
-    },
-    "thumb": {
-      "width": 10,
-      "height": 10
-    }
-  },
-  "variations": [
-    {
-      "id": "first",
-      "name": "First",
-      "meta": {
-        "desktop_screenshot": "desktop_light.jpg",
-        "mobile_screenshot": "mobile_light.jpg",
-        "description": "First variation",
-        "demo_url": "https://stencil-first.example.com",
-        "optimized_for": [
-          "arts_crafts"
-        ],
-        "industries": []
-      },
-      "settings": {
-        "color": "#000000",
-        "select": "first"
-      },
-      "images": {}
-    },
-    {
-      "id": "second",
-      "name": "Second",
-      "meta": {
-        "desktop_screenshot": "desktop_bold.jpg",
-        "mobile_screenshot": "mobile_bold.jpg",
-        "description": "Second variation",
-        "demo_url": "https://stencil-second.example.com",
-        "optimized_for": [
-          "arts_crafts"
-        ],
-        "industries": []
-      },
-      "settings": {
-        "color": "#000000",
-        "select": "second"
-      },
-      "images": {
-        "logo": {
-          "width": 200,
-          "height": 200
-        }
-      }
-    },
-    {
-      "id": "third",
-      "name": "Third",
-      "meta": {
-        "desktop_screenshot": "desktop_warm.jpg",
-        "mobile_screenshot": "mobile_warm.jpg",
-        "description": "Third variation",
-        "demo_url": "https://stencil-third.example.com",
-        "optimized_for": [
-          "arts_crafts"
-        ],
-        "industries": []
-      },
-      "settings": {
-        "color": "#FFFFFF",
-        "select": "third"
-      },
-      "images": {
-        "logo": {
-          "width": 150,
-          "height": 150
-        }
-      }
-    }
-  ]
-}
