--- test/_mocks/themes/invalid-scss-latest-node-sass-to-fix/config.stencil.json
@@ -1,10 +1,0 @@
-{
-  "customLayouts": {
-    "brand": {},
-    "category": {},
-    "page": {},
-    "product": {}
-  },
-  "normalStoreUrl": "https://url-from-answers.mybigcommerce.com",
-  "port": 3003
-}
