--- test/_mocks/themes/invalid-scss-latest-node-sass-to-fix/lang/en.json
@@ -1,20 +1,0 @@
-{
-  "header": {
-    "welcome_back": "Welcome back, {name}"
-  },
-  "footer": {
-    "brands": "Popular Brands",
-    "navigate": "Navigate",
-    "info": "Info",
-    "categories": "Categories",
-    "call_us": "Call us at {phone_number}"
-  },
-  "home": {
-    "heading": "Home"
-  },
-  "blog": {
-    "recent_posts": "Recent Posts",
-    "label": "Blog",
-    "posted_by": "Posted by {name}"
-  }
-}
