--- test/_mocks/themes/invalid-scss-latest-node-sass-to-fix/schema.json
@@ -1,103 +1,0 @@
-[
-  {
-    "name": "i18n.Test",
-    "settings": [
-      {
-        "type": "heading",
-        "content": "Heading"
-      },
-      {
-        "type": "paragraph",
-        "content": "Paragraph"
-      },
-      {
-        "type": "color",
-        "label": "Color",
-        "id": "color"
-      },
-      {
-        "type": "font",
-        "label": "Font",
-        "id": "font",
-        "options": [
-          {
-            "group": "Karla",
-            "label": "Karla",
-            "value": "Google_Karla_400"
-          },
-          {
-            "group": "Roboto",
-            "label": "Roboto",
-            "value": "Google_Roboto_400"
-          },
-          {
-            "group": "Source Sans Pro",
-            "label": "Source Sans Pro",
-            "value": "Google_Source+Sans+Pro_400"
-          }
-        ]
-      },
-      {
-        "type": "select",
-        "label": "Select",
-        "id": "select",
-        "force_reload": true,
-        "options": [
-          {
-            "value": "first",
-            "label": "First"
-          },
-          {
-            "value": "second",
-            "label": "Second"
-          },
-          {
-            "value": "third",
-            "label": "Third"
-          }
-        ]
-      },
-      {
-        "type": "checkbox",
-        "label": "Checkbox",
-        "id": "checkbox"
-      },
-      {
-        "type": "select",
-        "label": "Select",
-        "id": "select-two",
-        "options": [
-          {
-            "value": "first",
-            "label": "First"
-          },
-          {
-            "value": "second",
-            "label": "Second"
-          }
-        ]
-      }
-    ]
-  },
-  {
-    "name": "Page",
-    "settings": [
-      {
-        "type": "paragraph",
-        "label": "A Title",
-        "id": "customizable_title",
-        "content": "hello world"
-      },
-      {
-        "type": "checkbox",
-        "label": "Fetch Data",
-        "id": "front_matter_value"
-      },
-      {
-        "type": "input",
-        "label": "Display That",
-        "id": "display_that"
-      }
-    ]
-  }
-]
