--- test/_mocks/themes/invalid-scss-latest-node-sass-to-fix/schemaTranslations.json
@@ -1,5 +1,0 @@
-{
-  "i18n.Test": {
-    "default": "Test"
-  }
-}
