--- test/_mocks/themes/invalid-scss-latest-node-sass-to-fix/secrets.stencil.json
@@ -1,3 +1,0 @@
-{
-  "accessToken": "accessToken_from_answers"
-}
