--- test/_mocks/themes/invalid-scss-latest-node-sass-to-fix/templates/components/a.html
@@ -1,1 +1,0 @@
-a
