--- test/_mocks/themes/invalid-scss-latest-node-sass-to-fix/templates/components/b.html
@@ -1,1 +1,0 @@
-b
