--- test/_mocks/themes/invalid-scss-latest-node-sass-to-fix/templates/pages/page.html
@@ -1,19 +1,0 @@
----
-front_matter_options:
-    setting_x: {{theme_settings.front_matter_value}}
----
-<!DOCTYPE html>
-<html>
-<head>
-    <title>page.html</title>
-    {{head.scripts}}
-    {{ stylesheet 'assets/css/theme.css' }}
-</head>
-<body>
-    {{#if theme_settings.display_that}}
-        <div>That</div>
-        {{> components/a}}
-    {{/if}}
-    {{footer.scripts}}
-</body>
-</html>
