--- test/_mocks/themes/invalid-scss-latest-node-sass-to-fix/templates/pages/page2.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <title>page2.html</title>
-    {{head.scripts}}
-</head>
-<body>
-    <h1>{{theme_settings.customizable_title}}</h1>
-    {{> components/b}}
-    {{footer.scripts}}
-</body>
-</html>
